impls rating histogram and average rating widget 	loads them when viewing an apartment and updates them at real time if rating is changed

+++ app/src/main/java/com/csci5115/group8/data/Review.java
@@ -9,9 +9,9 @@ import java.util.Map;
 public class Review {
     public User user;
     public Apartment apartment;
-    public float rating;
+    public int rating;
 
-    public Review(User user, Apartment apartment, float rating) {
+    public Review(User user, Apartment apartment, int rating) {
         this.user = user;
         this.apartment = apartment;
         this.rating = rating;
